Move convertToWebSocketUrl to Loki
